Add response buffer to Ui
Currently, if there are storage errors, or if no existing storage file is found, we immediately try to send a status message indicating as such.

However, since no MainWindow has been created at that point in the program's lifecycle, this leads to a NullPointerException and causes the app to crash at launch.

Let's buffer responses if there is no mainWindow instance present, and add a check to respond() to print out the messages if there are buffered responses.

Since we always call a respond() a second time to show the welcome banner, this will not cause a delay in the status message being sent.